modelx: export ComputeSnapSize
This is an internal implementation detail currently used by x/porcelain. We
are planning on forking x/porcelain in probe-engine. So we cannot use any
internal package inside its implementation.
